server/requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
openpyxl>=3.1.0
fastapi>=0.95.0
uvicorn>=0.23.0
